Assigning ownership for carbon user
tendrl-bug-id:Tendrl/monitoring-integration#595
bugzilla: 1647322
