+++ tendrl/monitoring_integration/graphite/graphite_utils.py
@@ -1,10 +1,36 @@
 import ConfigParser
 import copy
+import grp
 import os
+import pwd
 import shutil
 
 from tendrl.commons.utils import log_utils as logger
 from tendrl.monitoring_integration.grafana import constants
+
+
+def change_owner(path, uname, gname, recursive=False):
+    try:
+        uid = pwd.getpwnam(uname).pw_uid
+        gid = grp.getgrnam(gname).gr_gid
+        _chown(path, uid, gid, recursive)
+    except KeyError as ex:
+        print (
+            "\n Unable to modify ownership of file/directory. "
+            "Error: %s \n" % ex
+        )
+
+
+def _chown(path, uid, gid, recursive):
+    os.chown(path, uid, gid)
+    if recursive:
+        for item in os.listdir(path):
+            itempath = os.path.join(path, item)
+            if os.path.isfile(itempath):
+                os.chown(itempath, uid, gid)
+            elif os.path.isdir(itempath):
+                os.chown(itempath, uid, gid)
+                _chown(itempath, uid, gid, recursive)
 
 
 def archive(
